requirements.txt
setuptools~=72.1.0
pillow~=10.4.0
datasets~=2.20.0
torchvision~=0.19.0
numpy~=1.26.4
accelerate~=0.33.0
diffusers~=0.30.0
transformers~=4.44.0
tqdm~=4.66.4
bitsandbytes